Move ZcBuf to separate files.
